savepoint

+++ src/util.rs
@@ -6,7 +6,8 @@ pub fn frac_to_time(frac: &fraction::prelude::GenericFraction<u64>) -> Time {
     Time {
         num: (*frac.fract().numer().unwrap_or(&0)
             + frac.trunc().numer().unwrap_or(&0) * *frac.fract().denom().unwrap_or(&0))
-            as usize,
+            as isize
+            * (if frac.is_sign_positive() { 1 } else { -1 }),
         denom: *frac.fract().denom().unwrap_or(&0) as usize,
     }
 }
@@ -17,7 +18,7 @@ pub fn time_to_frac(time: &Time) -> fraction::prelude::GenericFraction<u64> {
 
 pub fn time_expression_to_time(time_expression: &TimeExpression) -> Time {
     Time {
-        num: time_expression.beat * 4 + time_expression.sixteenth.unwrap_or(0),
+        num: (time_expression.beat * 4 + time_expression.sixteenth.unwrap_or(0)) as isize,
         denom: 16,
     }
 }
